backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Vector Database
chromadb==0.4.18

# AI & ML
langchain==0.1.0
openai==1.6.1
sentence-transformers==2.2.2
transformers==4.36.0
dashscope==1.10.0
erniebot==0.5.0
numpy==1.26.2

# Document Processing
python-docx==1.1.0
pypdf==5.1.0
pdfplumber==0.10.3
openpyxl==3.1.2
python-multipart==0.0.6
jinja2==3.1.2
markdown2==2.4.10

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Utilities
httpx==0.25.2
aiofiles==23.2.1
python-dateutil==2.8.2

# Redis
redis==5.0.1
hiredis==2.2.3

# Logging & Monitoring
loguru==0.7.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
